requirements.txt
pytest~=7.1.1
python-dotenv~=0.20.0
SQLAlchemy~=1.4.39
psycopg2==2.8.6